// src/app/core/api/v1/model/investment.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Style } from './style';


export interface Investment { 
    id?: number;
    name?: string;
    fullyVetted?: boolean;
    price?: number;
    style?: Style;
}